feat: refactor `HorizontalScrollFade` into a separate file
